+++ unittest_binance_manager.py
@@ -37,75 +37,202 @@ from unicorn_binance_websocket_api.unicorn_binance_websocket_api_manager import 
 import unittest
 import uuid
 
+BINANCE_COM_API_KEY = ""
+BINANCE_COM_API_SECRET = ""
+BINANCE_JE_API_KEY = ""
+BINANCE_JE_API_SECRET = ""
 
-class TestBinanceManager(unittest.TestCase):
+
+class TestBinanceComManager(unittest.TestCase):
+    # Test binance.com (Binance)
 
     def setUp(self):
-        self.binance_com_api_key = ""
-        self.binance_com_api_secret = ""
-
-        self.binance_websocket_api_manager_com = BinanceWebSocketApiManager(exchange="binance.com")
-        self.binance_websocket_api_manager_je = BinanceWebSocketApiManager(exchange="binance.je")
-        self.binance_websocket_api_manager_org = BinanceWebSocketApiManager(exchange="binance.org")
-        self.binance_websocket_api_manager_org_testnet = BinanceWebSocketApiManager(exchange="binance.org-testnet")
+        self.binance_com_api_key = BINANCE_COM_API_KEY
+        self.binance_com_api_secret = BINANCE_COM_API_SECRET
+        self.binance_com_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.com")
 
     def test_create_uri_miniticker_regular_com(self):
-        self.assertEqual(self.binance_websocket_api_manager_com.create_websocket_uri(["!miniTicker"], ["arr"]),
+        self.assertEqual(self.binance_com_websocket_api_manager.create_websocket_uri(["!miniTicker"], ["arr"]),
                          'wss://stream.binance.com:9443/ws/!miniTicker@arr')
 
     def test_create_uri_miniticker_reverse_com(self):
-        self.assertEqual(self.binance_websocket_api_manager_com.create_websocket_uri(["arr"], ["!miniTicker"]),
+        self.assertEqual(self.binance_com_websocket_api_manager.create_websocket_uri(["arr"], ["!miniTicker"]),
                          'wss://stream.binance.com:9443/ws/!miniTicker@arr')
 
     def test_create_uri_ticker_regular_com(self):
-        self.assertEqual(self.binance_websocket_api_manager_com.create_websocket_uri(["!ticker"], ["arr"]),
+        self.assertEqual(self.binance_com_websocket_api_manager.create_websocket_uri(["!ticker"], ["arr"]),
                          'wss://stream.binance.com:9443/ws/!ticker@arr')
 
     def test_create_uri_ticker_reverse_com(self):
-        self.assertEqual(self.binance_websocket_api_manager_com.create_websocket_uri(["arr"], ["!ticker"]),
+        self.assertEqual(self.binance_com_websocket_api_manager.create_websocket_uri(["arr"], ["!ticker"]),
                          'wss://stream.binance.com:9443/ws/!ticker@arr')
 
-    def test_create_uri_userdata_regular_com_false(self):
-        self.assertFalse(self.binance_websocket_api_manager_com.create_websocket_uri(["!userData"], ["arr"]))
+    def test_create_uri_userdata_regular_false_com(self):
+        self.assertFalse(self.binance_com_websocket_api_manager.create_websocket_uri(["!userData"], ["arr"]))
 
-    def test_create_uri_userdata_reverse_com_false(self):
-        self.assertFalse(self.binance_websocket_api_manager_com.create_websocket_uri(["arr"], ["!userData"]))
+    def test_create_uri_userdata_reverse_false_com(self):
+        self.assertFalse(self.binance_com_websocket_api_manager.create_websocket_uri(["arr"], ["!userData"]))
 
     def test_create_uri_userdata_regular_com(self):
+        if len(self.binance_com_api_key) == 0 or len(self.binance_com_api_secret) == 0:
+            print("\r\nempty API key and/or secret: can not successfully test test_create_uri_userdata_regular_com() "
+                  "for binance.com")
         stream_id = uuid.uuid4()
-        self.binance_websocket_api_manager_com._add_socket_to_socket_list(stream_id, ["!userData"], ["arr"])
-        self.assertRegex(self.binance_websocket_api_manager_com.create_websocket_uri(["!userData"], ["arr"],
+        self.binance_com_websocket_api_manager._add_socket_to_socket_list(stream_id, ["!userData"], ["arr"])
+        self.assertRegex(self.binance_com_websocket_api_manager.create_websocket_uri(["!userData"], ["arr"],
                                                                                      stream_id,
                                                                                      self.binance_com_api_key,
                                                                                      self.binance_com_api_secret),
                          r'wss://stream.binance.com:9443/ws/.')
 
     def test_create_uri_userdata_reverse_com(self):
+        if len(self.binance_com_api_key) == 0 or len(self.binance_com_api_secret) == 0:
+            print("\r\nempty API key and/or secret: can not successfully test test_create_uri_userdata_reverse_com() "
+                  "for binance.com")
         stream_id = uuid.uuid4()
-        self.binance_websocket_api_manager_com._add_socket_to_socket_list(stream_id, ["arr"], ["!userData"])
-        self.assertRegex(self.binance_websocket_api_manager_com.create_websocket_uri(["arr"], ["!userData"],
+        self.binance_com_websocket_api_manager._add_socket_to_socket_list(stream_id, ["arr"], ["!userData"])
+        self.assertRegex(self.binance_com_websocket_api_manager.create_websocket_uri(["arr"], ["!userData"],
                                                                                      stream_id,
                                                                                      self.binance_com_api_key,
                                                                                      self.binance_com_api_secret),
                          r'wss://stream.binance.com:9443/ws/.')
 
-    def test_create_uri_single_regular_com(self):
-        self.assertEqual(self.binance_websocket_api_manager_com.create_websocket_uri(["trade"], ["bnbbtc"]),
+    def test_create_uri_single_com(self):
+        self.assertEqual(self.binance_com_websocket_api_manager.create_websocket_uri(["trade"], ["bnbbtc"]),
                          'wss://stream.binance.com:9443/stream?streams=bnbbtc@trade/')
 
-    def test_create_uri_multi_regular_com(self):
-        self.assertEqual(self.binance_websocket_api_manager_com.create_websocket_uri(['trade', 'kline_1'],
+    def test_create_uri_multi_com(self):
+        self.assertEqual(self.binance_com_websocket_api_manager.create_websocket_uri(['trade', 'kline_1'],
                                                                                      ['bnbbtc', 'ethbtc']),
                          'wss://stream.binance.com:9443/stream?streams=bnbbtc@trade/ethbtc@trade/'
                          'bnbbtc@kline_1/ethbtc@kline_1/')
 
     def tearDown(self):
-        self.binance_websocket_api_manager_com.stop_manager_with_all_streams()
-        self.binance_websocket_api_manager_je.stop_manager_with_all_streams()
-        self.binance_websocket_api_manager_org.stop_manager_with_all_streams()
-        self.binance_websocket_api_manager_org_testnet.stop_manager_with_all_streams()
+        self.binance_com_websocket_api_manager.stop_manager_with_all_streams()
 
 
+class TestBinanceJeManager(unittest.TestCase):
+    # Test binance.je (Binance Jersey)
+
+    def setUp(self):
+        self.binance_je_api_key = BINANCE_JE_API_KEY
+        self.binance_je_api_secret = BINANCE_JE_API_SECRET
+
+        self.binance_je_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.je")
+
+    def test_create_uri_miniticker_regular_je(self):
+        self.assertEqual(self.binance_je_websocket_api_manager.create_websocket_uri(["!miniTicker"], ["arr"]),
+                         'wss://stream.binance.je:9443/ws/!miniTicker@arr')
+
+    def test_create_uri_miniticker_reverse_je(self):
+        self.assertEqual(self.binance_je_websocket_api_manager.create_websocket_uri(["arr"], ["!miniTicker"]),
+                         'wss://stream.binance.je:9443/ws/!miniTicker@arr')
+
+    def test_create_uri_ticker_regular_je(self):
+        self.assertEqual(self.binance_je_websocket_api_manager.create_websocket_uri(["!ticker"], ["arr"]),
+                         'wss://stream.binance.je:9443/ws/!ticker@arr')
+
+    def test_create_uri_ticker_reverse_je(self):
+        self.assertEqual(self.binance_je_websocket_api_manager.create_websocket_uri(["arr"], ["!ticker"]),
+                         'wss://stream.binance.je:9443/ws/!ticker@arr')
+
+    def test_create_uri_userdata_regular_false_je(self):
+        self.assertFalse(self.binance_je_websocket_api_manager.create_websocket_uri(["!userData"], ["arr"]))
+
+    def test_create_uri_userdata_reverse_false_je(self):
+        self.assertFalse(self.binance_je_websocket_api_manager.create_websocket_uri(["arr"], ["!userData"]))
+
+    def test_create_uri_userdata_regular_je(self):
+        if len(self.binance_je_api_key) == 0 or len(self.binance_je_api_secret) == 0:
+            print("\r\nempty API key and/or secret: can not successfully test test_create_uri_userdata_regular_je() "
+                  "for binance.com")
+        stream_id = uuid.uuid4()
+        self.binance_je_websocket_api_manager._add_socket_to_socket_list(stream_id, ["!userData"], ["arr"])
+        self.assertRegex(self.binance_je_websocket_api_manager.create_websocket_uri(["!userData"], ["arr"],
+                                                                                     stream_id,
+                                                                                     self.binance_je_api_key,
+                                                                                     self.binance_je_api_secret),
+                         r'wss://stream.binance.je:9443/ws/.')
+
+    def test_create_uri_userdata_reverse_je(self):
+        if len(self.binance_je_api_key) == 0 or len(self.binance_je_api_secret) == 0:
+            print("\r\nempty API key and/or secret: can not successfully test test_create_uri_userdata_reverse_je() "
+                  "for binance.com")
+        stream_id = uuid.uuid4()
+        self.binance_je_websocket_api_manager._add_socket_to_socket_list(stream_id, ["arr"], ["!userData"])
+        self.assertRegex(self.binance_je_websocket_api_manager.create_websocket_uri(["arr"], ["!userData"],
+                                                                                     stream_id,
+                                                                                     self.binance_je_api_key,
+                                                                                     self.binance_je_api_secret),
+                         r'wss://stream.binance.je:9443/ws/.')
+
+    def test_create_uri_single_je(self):
+        self.assertEqual(self.binance_je_websocket_api_manager.create_websocket_uri(["trade"], ["bnbbtc"]),
+                         'wss://stream.binance.je:9443/stream?streams=bnbbtc@trade/')
+
+    def test_create_uri_multi_je(self):
+        self.assertEqual(self.binance_je_websocket_api_manager.create_websocket_uri(['trade', 'kline_1'],
+                                                                                     ['bnbbtc', 'ethbtc']),
+                         'wss://stream.binance.je:9443/stream?streams=bnbbtc@trade/ethbtc@trade/'
+                         'bnbbtc@kline_1/ethbtc@kline_1/')
+
+    def tearDown(self):
+        self.binance_je_websocket_api_manager.stop_manager_with_all_streams()
+
+
+class TestBinanceOrgManager(unittest.TestCase):
+    # Test binance.org (Binance Chain Dex)
+
+    def setUp(self):
+        self.binance_org_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.org")
+
+    def test_create_uri_alltickers_regular_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["$all"], ["allTickers"]),
+                         'wss://dex.binance.org/api/ws/$all@allTickers')
+
+    def test_create_uri_alltickers_reverse_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["allTickers"], ["$all"]),
+                         'wss://dex.binance.org/api/ws/$all@allTickers')
+
+    def test_create_uri_allminitickers_regular_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["$all"], ["allMiniTickers"]),
+                         'wss://dex.binance.org/api/ws/$all@allMiniTickers')
+
+    def test_create_uri_allminitickers_reverse_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["allMiniTickers"], ["$all"]),
+                         'wss://dex.binance.org/api/ws/$all@allMiniTickers')
+
+    def test_create_uri_blockheight_regular_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["$all"], ["blockheight"]),
+                         'wss://dex.binance.org/api/ws/$all@blockheight')
+
+    def test_create_uri_blockheight_reverse_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["blockheight"], ["$all"]),
+                         'wss://dex.binance.org/api/ws/$all@blockheight')
+
+    def test_create_uri_single_trades_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["trades"], ["RAVEN-F66_BNB"]),
+                         'wss://dex.binance.org/api/ws/RAVEN-F66_BNB@trades')
+
+    def test_create_uri_single_marketdepth_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["marketDepth"], ["RAVEN-F66_BNB"]),
+                         'wss://dex.binance.org/api/ws/RAVEN-F66_BNB@marketDepth')
+
+    def test_create_uri_single_kline_1h_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["kline_1h"], ["RAVEN-F66_BNB"]),
+                         'wss://dex.binance.org/api/ws/RAVEN-F66_BNB@kline_1h')
+
+    def test_create_uri_single_ticker_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["ticker"], ["RAVEN-F66_BNB"]),
+                         'wss://dex.binance.org/api/ws/RAVEN-F66_BNB@ticker')
+
+    def test_create_uri_single_miniTicker_org(self):
+        self.assertEqual(self.binance_org_websocket_api_manager.create_websocket_uri(["miniTicker"], ["RAVEN-F66_BNB"]),
+                         'wss://dex.binance.org/api/ws/RAVEN-F66_BNB@miniTicker')
+
+
+    def tearDown(self):
+        self.binance_org_websocket_api_manager.stop_manager_with_all_streams()
 
 
 if __name__ == '__main__':